requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
openai==1.3.0
pandas
streamlit
altair>=4.0.0
langgraph==0.0.50
python-dotenv